requirements.txt
alembic==1.7.7
async-timeout==4.0.2
cffi==1.15.0
click==8.1.0
cryptography==36.0.2
Deprecated==1.2.13
Flask-Caching==1.10.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask==2.0.3
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
MyApplication==0.1.0
packaging==21.3
pip==22.0.4
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
redis==4.2.0
setuptools==61.2.0
six==1.16.0
SQLAlchemy==1.4.32
typing-extensions==4.1.1
Werkzeug==2.1.0
wheel==0.37.1
wrapt==1.14.0
pydevd-pycharm~=213.7172.26